tet: Fix the remaining warnings.
Add a TET_NORETURN attribute and tag various fatal functions with it.  That
fixes a few "int x; switch(...) { case C: x = blah; break; default: fatal() }"
cases that result in uninitialized variable warnings.

Acked-by: Peter Hutterer <peter.hutterer@who-t.net>

+++ src/tet3/tcc/tool.c
@@ -660,7 +660,7 @@ int xrfnamelen;
 	const char *tcname = tet_basename(tcpath);
 	char logname[64];
 
-	sprintf(logname, "%.*s.log", sizeof(logname) - 5, tcname);
+	sprintf(logname, "%.*s.log", (int)(sizeof(logname) - 5), tcname);
 	tcdirfname(tcpath, logname, xrfname, xrfnamelen);
 }
